MINOR: cfgparse: Add a cast to make gcc happier.
When calling calloc(), cast global.nbthread to unsigned int, so that gcc
doesn't freak out, as it has no way of knowing global.nbthread can't be
negative.

(cherry picked from commit dd1c8f1f721fc139ecb5b96f81caf69894964715)
[wt: adjusted context]

+++ src/cfgparse.c
@@ -3401,7 +3401,7 @@ out_uri_auth_compat:
 			}
 
 			if (newsrv->max_idle_conns != 0) {
-				newsrv->idle_orphan_conns = calloc(global.nbthread, sizeof(*newsrv->idle_orphan_conns));
+				newsrv->idle_orphan_conns = calloc((unsigned int)global.nbthread, sizeof(*newsrv->idle_orphan_conns));
 				newsrv->idle_task         = calloc(global.nbthread, sizeof(*newsrv->idle_task));
 				if (!newsrv->idle_orphan_conns || !newsrv->idle_task)
 					goto err;
